requirements.txt
python-dotenv
google-cloud-firestore
fastapi
uvicorn
